Geo Location
Javascript -- seems like javascript solution wont work as javascript is
client side and sinatra is server side. I think we would need to use
asp.net or PHP somehow so that they can be processed on the server. I am
not very experianced with these technologies. It looks like the slim gem
could potentially have a solution but I tried looking at how to use it
and am confused.